requirements.txt
pandas~=0.24.2
singer-python~=5.10.0
tap-salesforce~=1.5.5
simple_salesforce
requests~=2.21.0
python_decouple
simple_salesforce
logging
tap-salesforce
target-csv
# singer-discover
gluestick
python-dotenv~=0.20.0
# python~=3.7.11
datetime